Filling out response types
Also use a singleton for remembering the "loader chain", because otherwise every implementor of `HTTPLoader` has to do cycle detection and provide storage for the next loader.

But this also means that the singleton has to keep a strong reference to every loader ever used in a chain, which makes it hard to destroy a chain if you're done with it.