URLSessionLoader
It would be pretty simple to make a URLSession be an HTTPLoader by just using the normal async methods on URLSession. However, those methods don't offer a way to interact with the delegate methods of a URLSession, which is where customization around redirection, authentication challenges, progress reporting, etc happen.

So, this implements a custom async interface for URLSession by starting a continuation and the holding on to it until the task finishes. This means we can use the delegate methods to interact with the request and any options it has, all from a "single" async call.